Fix scope and data type issues in model_architecture.py

+++ scripts/model_architecture.py
@@ -12,7 +12,7 @@ class VishwamAIModel(hk.Module):
         self.tokenizer.pad_token = self.tokenizer.eos_token  # Set padding token to eos token
         self.transformer = hk.transform(lambda x: hk.Sequential([
             hk.Embed(vocab_size=50257, embed_dim=512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", "uniform")),
-            lambda x: self.attention(query=x, key=x, value=x),
+            lambda x: self.attention(query=x.astype(jnp.int32), key=x.astype(jnp.int32), value=x.astype(jnp.int32)),
             hk.Linear(2048, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", "uniform")),
             hk.Linear(512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", "uniform"))
         ])(x.astype(jnp.int32)))
@@ -32,7 +32,7 @@ class VishwamAIModel(hk.Module):
         self.num_experts = 4  # Reduced number of experts to 4
         self.experts = [hk.transform(lambda x: hk.Sequential([
             hk.Embed(vocab_size=50257, embed_dim=512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", "uniform")),
-            lambda x: self.attention(query=x, key=x, value=x),
+            lambda x: self.attention(query=x.astype(jnp.int32), key=x.astype(jnp.int32), value=x.astype(jnp.int32)),
             hk.Linear(2048, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", "uniform")),
             hk.Linear(512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", "uniform"))
         ])(x.astype(jnp.int32))) for _ in range(self.num_experts)]
